# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "herzen"
  },
  "paths": {
...
    "/api/v1/gto": {
      "get": {
        "tags": [
          "GTO"
        ],
        "summary": "Get GTO by institute id",
        "description": "get data of GTO by institute id",
        "operationId": "get-api-v1-gto",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer Token",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "institute_id",
            "in": "query",
            "description": "id of institute (0:infinity)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "student_id": {
                      "type": "integer"
                    },
                    "level": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "student_id": 0,
                      "level": "string"
                    }
                  }
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "count_by_institute": {
                        "gold": 0,
                        "silver": 0,
                        "bronze": 0
                      },
                      "members_count": {
                        "gold": 0,
                        "silver": 0,
                        "bronze": 0
                      },
                      "rating": {
                        "gold": 0,
                        "silver": 0,
                        "bronze": 0
                      },
                      "percent_by_common": {
                        "gold": 0,
                        "silver": 0,
                        "bronze": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "",
            "description": ""
          }
        ]
      },
      "patch": {
        "tags": [
          "GTO"
        ],
        "summary": "PATCH GTO results",
        "operationId": "patch-api-v1-gto",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer Token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "{\n    \"student_id\": 0,\n    \"level\": \"string\"\n}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "student_id": {
                    "type": "integer"
                  },
                  "level": {
                    "type": "string"
                  }
                },
                "x-examples": {
                  "Example 1": {
                    "student_id": 0,
                    "level": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "student_id": {
                      "type": "integer"
                    },
                    "level": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "student_id": 0,
                      "level": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "",
            "description": ""
          }
        ]
      },
      "post": {
        "tags": [
          "GTO"
        ],
        "summary": "POST GTO results",
        "operationId": "post-api-v1-gto",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "{\n    \"student_id\": 0,\n    \"level\": \"string\"\n}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "student_id": {
                    "type": "integer"
                  },
                  "level": {
                    "type": "string"
                  }
                },
                "x-examples": {
                  "Example 1": {
                    "student_id": 0,
                    "level": "string"
                  }
                }
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "student_id": 0,
                    "level": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "student_id": {
                      "type": "integer"
                    },
                    "level": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "student_id": 0,
                      "level": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "msg": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "detail": [
                        {
                          "loc": [
                            "string",
                            0
                          ],
                          "msg": "string",
                          "type": "string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "",
            "description": ""
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/api/v1/login/access-token",
            "scopes": {}
          }
        }
      }
    },
...
  }
}